@@ -91,12 +91,12 @@ def test_edge_cases(df):
     with pytest.raises(ValueError) as error:
         ln.Collection(df, invalid_param=1)
     assert str(error.exconly()).startswith(
-        "ValueError: Only artifacts, name, run, description, reference, reference_type, visibility can be passed, you passed: "
+        "ValueError: Only artifacts, key, run, description, reference, reference_type can be passed, you passed: "
     )
     with pytest.raises(ValueError) as error:
         ln.Collection(1, name="Invalid")
     assert str(error.exconly()).startswith(
-        "ValueError: Artifact or List[Artifact] is allowed."
+        "ValueError: Artifact or list[Artifact] is allowed."
     )
     artifact = ln.Artifact.from_df(df, description="Test artifact")
     assert artifact._state.adding
@@ -116,12 +116,9 @@ def test_edge_cases(df):
 
 
 def test_from_inconsistent_artifacts(df, adata):
-    artifact1 = ln.Artifact.from_df(df, description="My test")
-    artifact1.save()
-    artifact2 = ln.Artifact.from_anndata(adata, description="My test2")
-    artifact2.save()
-    collection = ln.Collection([artifact1, artifact2], name="Inconsistent")
-    collection.save()
+    artifact1 = ln.Artifact.from_df(df, description="My test").save()
+    artifact2 = ln.Artifact.from_anndata(adata, description="My test2").save()
+    collection = ln.Collection([artifact1, artifact2], name="Inconsistent").save()
     # test idempotency of .save()
     collection.save()
     # create a run context
@@ -174,7 +171,7 @@ def test_from_consistent_artifacts(adata, adata2):
     collection2 = ln.Collection([artifact1, artifact2], name="My test 1", run=run)
     assert not collection2._state.adding
     assert collection2.id == collection.id
-    assert collection2.name == "My test 1"
+    assert collection2.key == "My test 1"
 
     collection.delete(permanent=True)
     artifact1.delete(permanent=True)
@@ -379,15 +376,13 @@ def test_collection_mapped(adata, adata2):
 
 def test_revise_collection(df, adata):
     # create a versioned collection
-    artifact = ln.Artifact.from_df(df, description="test")
-    artifact.save()
+    artifact = ln.Artifact.from_df(df, description="test").save()
     collection = ln.Collection(artifact, name="test", version="1")
     assert collection.version == "1"
     assert collection.uid.endswith("0000")
     collection.save()
 
-    artifact = ln.Artifact.from_anndata(adata, description="test")
-    artifact.save()
+    artifact = ln.Artifact.from_anndata(adata, description="test").save()
 
     with pytest.raises(ValueError) as error:
         collection_r2 = ln.Collection(artifact, revises=collection, version="1")
@@ -397,14 +392,14 @@ def test_revise_collection(df, adata):
         ln.Collection(adata, revises="wrong-type")
 
     # create new collection from old collection
-    collection_r2 = ln.Collection(artifact, is_new_version_of=collection)
+    collection_r2 = ln.Collection(artifact, revises=collection)
     assert collection_r2.stem_uid == collection.stem_uid
     assert collection_r2.uid.endswith("0001")
     collection_r2 = ln.Collection(artifact, revises=collection)
     assert collection_r2.stem_uid == collection.stem_uid
     assert collection_r2.uid.endswith("0001")
     assert collection_r2.version is None
-    assert collection_r2.name == "test"
+    assert collection_r2.key == "test"
 
     collection_r2.save()
 
@@ -418,7 +413,7 @@ def test_revise_collection(df, adata):
     assert collection_r3.stem_uid == collection.stem_uid
     assert collection_r3.version == "2"
     assert collection_r3.uid.endswith("0002")
-    assert collection_r3.name == "test1"
+    assert collection_r3.key == "test1"
 
     artifacts_r2 = collection_r2.artifacts.all()
     collection_r2.delete(permanent=True)
@@ -429,13 +424,9 @@ def test_revise_collection(df, adata):
 
 
 def test_collection_append(df, adata):
-    artifact = ln.Artifact.from_df(df, description="test")
-    artifact.save()
-    artifact_1 = ln.Artifact.from_anndata(adata, description="test")
-    artifact_1.save()
-
-    col = ln.Collection(artifact, name="Test", description="Test append")
-    col.save()
+    artifact = ln.Artifact.from_df(df, description="test").save()
+    artifact_1 = ln.Artifact.from_anndata(adata, description="test").save()
+    col = ln.Collection(artifact, name="Test", description="Test append").save()
 
     col_append = col.append(artifact_1).save()
 
