@@ -832,11 +832,11 @@ def _add_values(
     model_name = "Param" if is_param else "Feature"
     if is_param:
         if self._host.__class__ == Artifact:
-            if self._host.type != "model":
+            if self._host.kind != "model":
                 raise ValidationError("Can only set params for model-like artifacts.")
     else:
         if self._host.__class__ == Artifact:
-            if self._host.type != "dataset" and self._host.type is not None:
+            if self._host.kind != "dataset" and self._host.kind is not None:
                 raise ValidationError(
                     "Can only set features for dataset-like artifacts."
                 )
@@ -1089,10 +1089,10 @@ def _add_set_from_df(
 ):
     """Add feature set corresponding to column names of DataFrame."""
     if isinstance(self._host, Artifact):
-        assert self._host._accessor == "DataFrame"  # noqa: S101
+        assert self._host.otype == "DataFrame"  # noqa: S101
     else:
         # Collection
-        assert self._host.artifact._accessor == "DataFrame"  # noqa: S101
+        assert self._host.artifact.otype == "DataFrame"  # noqa: S101
     df = self._host.load()
     feature_set = FeatureSet.from_df(
         df=df,
@@ -1113,7 +1113,7 @@ def _add_set_from_anndata(
 ):
     """Add features from AnnData."""
     if isinstance(self._host, Artifact):
-        assert self._host._accessor == "AnnData"  # noqa: S101
+        assert self._host.otype == "AnnData"  # noqa: S101
     else:
         raise NotImplementedError()
 
@@ -1143,7 +1143,7 @@ def _add_set_from_mudata(
     if obs_fields is None:
         obs_fields = {}
     if isinstance(self._host, Artifact):
-        assert self._host._accessor == "MuData"  # noqa: S101
+        assert self._host.otype == "MuData"  # noqa: S101
     else:
         raise NotImplementedError()
 
