@@ -133,15 +133,18 @@ def get_new_path_from_uid(old_path: UPath, old_uid: str, new_uid: str):
 def process_revises(
     revises: IsVersioned | None,
     version: str | None,
-    name: str | None,
+    key: str | None,
+    description: str | None,
     type: type[IsVersioned],
-) -> tuple[str, str, str, IsVersioned | None]:
+) -> tuple[str, str, str, str, IsVersioned | None]:
     if revises is not None and not isinstance(revises, type):
         raise TypeError(f"`revises` has to be of type `{type.__name__}`")
     uid, revises = create_uid(
         revises=revises, version=version, n_full_id=type._len_full_uid
     )
     if revises is not None:
-        if name is None:
-            name = revises.name
-    return uid, version, name, revises
+        if description is None:
+            description = revises.description
+        if key is None:
+            key = revises.key
+    return uid, version, key, description, revises
