@@ -101,26 +101,26 @@ def test_create_or_load_transform():
     context.version = version
     context._path = Path("my-test-transform-create-or-load.py")
     context._create_or_load_transform(
-        name=title,
+        description=title,
         transform_type="notebook",
     )
     assert context._transform.uid == uid
     assert context._transform.version == version
-    assert context._transform.name == title
+    assert context._transform.description == title
     context._create_or_load_transform(
-        name=title,
+        description=title,
     )
     assert context._transform.uid == uid
     assert context._transform.version == version
-    assert context._transform.name == title
+    assert context._transform.description == title
 
     # now, test an updated transform name
     context._create_or_load_transform(
-        name="updated title",
+        description="updated title",
     )
     assert context._transform.uid == uid
     assert context._transform.version == version
-    assert context._transform.name == "updated title"
+    assert context._transform.description == "updated title"
 
 
 def test_run_scripts():
@@ -250,12 +250,11 @@ def test_run_external_script():
         "https://github.com/laminlabs/lamin-cli/blob/"
     )
     assert transform.reference_type == "url"
-    assert transform.name == "My good script"
+    assert transform.description == "My good script"
     # ensure that the source code is not saved as an output artifact
     assert transform.latest_run.output_artifacts.count() == 0
     assert transform.runs.count() == 1
     assert transform.hash == "MoIciBQ0lpVPCKQGofPX6g"
-    assert transform._source_code_artifact is None
 
 
 @pytest.mark.parametrize("type", ["notebook", "script"])
