@@ -25,13 +25,20 @@ def test_revise_transforms():
     # try to reload the same transform with the same uid
     transform_reload = ln.Transform(uid=transform.uid, name="My transform updated name")
     assert transform_reload.id == transform.id
-    assert transform_reload.name == "My transform updated name"
+    assert transform_reload.key == "My transform"  # unchanged, prints logging
+    transform_reload = ln.Transform(
+        uid=transform.uid, description="My transform updated name"
+    )
+    assert transform_reload.id == transform.id
+    assert (
+        transform_reload.description == "My transform updated name"
+    )  # unchanged, prints logging
 
     # create new transform from old transform
-    transform_r2 = ln.Transform(name="My 2nd transform", is_new_version_of=transform)
+    transform_r2 = ln.Transform(description="My 2nd transform", revises=transform)
     assert transform_r2.uid != transform.uid
     assert transform_r2.uid.endswith("0001")
-    transform_r2 = ln.Transform(name="My 2nd transform", revises=transform)
+    transform_r2 = ln.Transform(description="My 2nd transform", revises=transform)
     assert transform_r2.uid != transform.uid
     assert transform_r2.uid.endswith("0001")
     assert transform_r2.stem_uid == transform.stem_uid
@@ -42,20 +49,29 @@ def test_revise_transforms():
     assert not transform.is_latest
 
     # create new transform from newly versioned transform
-    transform_r3 = ln.Transform(name="My transform", revises=transform_r2, version="2")
+    transform_r3 = ln.Transform(
+        description="My transform", revises=transform_r2, version="2"
+    )
     assert transform_r3.stem_uid == transform.stem_uid
     assert transform_r3.version == "2"
 
-    # default name
+    # default description
     transform_r3 = ln.Transform(revises=transform_r2)
-    assert transform_r3.name == transform_r2.name
+    assert transform_r3.description == transform_r2.description
 
     # revise by matching on `key`
     key = "my-notebook.ipynb"
     transform_r2.key = key
     transform_r2.save()
     assert transform_r2.is_latest
-    transform_r3 = ln.Transform(name="My transform", key=key, version="2")
+    transform_r3 = ln.Transform(description="My transform", key=key, version="2")
+    assert transform_r3.uid[:-4] == transform_r2.uid[:-4]
+    assert transform_r3.uid.endswith("0001")
+    # this only fires if source code was actually saved
+    transform_r2.source_code = "something"
+    transform_r2.save()
+    transform_r3 = ln.Transform(description="My transform", key=key, version="2")
+    assert transform_r3.uid[:-4] == transform_r2.uid[:-4]
     assert transform_r3.uid.endswith("0002")
     assert transform_r3.stem_uid == transform_r2.stem_uid
     assert transform_r3.key == key
@@ -81,7 +97,7 @@ def test_revise_transforms():
     with pytest.raises(ValueError) as error:
         ln.Transform(x=1)
     assert (
-        error.exconly() == "ValueError: Only name, key, version, type, revises,"
+        error.exconly() == "ValueError: Only key, description, version, type, revises,"
         " reference, reference_type can be passed, but you passed: {'x': 1}"
     )
 
@@ -98,7 +114,7 @@ def test_revise_transforms():
     transform.save()
 
     # create new transform from old transform
-    new_transform = ln.Transform(name="My new transform", revises=transform)
+    new_transform = ln.Transform(description="My new transform", revises=transform)
     assert transform.version is None
     assert new_transform.stem_uid == transform.stem_uid
     assert new_transform.uid.endswith("0001")
@@ -109,51 +125,29 @@ def test_revise_transforms():
 
 def test_delete():
     # prepare the creation of a transform with its artifacts
-    transform = ln.Transform(name="My transform")
-    transform.save()
+    transform = ln.Transform(name="My transform").save()
     run = ln.Run(transform)
     report_path = Path("report.html")
     with open(report_path, "w") as f:
         f.write("a")
-    _source_code_artifact_path = Path("code.py")
-    with open(_source_code_artifact_path, "w") as f:
-        f.write("b")
     environment_path = Path("environment.txt")
     with open(environment_path, "w") as f:
         f.write("c")
-    report = ln.Artifact(report_path, description=f"Report of {run.uid}")
-    report.save()
+    report = ln.Artifact(report_path, description=f"Report of {run.uid}").save()
     report_path.unlink()
     report_path = report.path
-    _source_code_artifact = ln.Artifact(
-        _source_code_artifact_path, description=f"Source of {transform.uid}"
-    )
-    _source_code_artifact.save()
-    _source_code_artifact_path.unlink()
-    _source_code_artifact_path = _source_code_artifact.path
-    environment = ln.Artifact(environment_path, description="requirement.txt")
-    environment.save()
+    environment = ln.Artifact(environment_path, description="requirements.txt").save()
     environment_path.unlink()
     environment_path = environment.path
-    transform._source_code_artifact = _source_code_artifact
     transform.save()
     run.report = report
     run.environment = environment
     run.save()
     assert report_path.exists()
-    assert _source_code_artifact_path.exists()
     assert environment_path.exists()
     # now delete everything
     transform.delete()
     assert not report_path.exists()
-    assert not _source_code_artifact_path.exists()
     assert not environment_path.exists()
-    assert (
-        len(
-            ln.Artifact.filter(
-                id__in=[report.id, _source_code_artifact.id, environment.id]
-            ).all()
-        )
-        == 0
-    )
+    assert len(ln.Artifact.filter(id__in=[report.id, environment.id]).all()) == 0
     assert len(ln.Run.filter(id=run.id).all()) == 0
