@@ -88,37 +88,36 @@ def test_get_new_path_from_uid():
 
 def test_latest_version_and_get():
     # build one version family
-    transform_v1 = ln.Transform(name="Introduction")
-    transform_v1.save()
+    transform_v1 = ln.Transform(description="Introduction").save()
     assert transform_v1.is_latest
     assert transform_v1.version is None
     # pass the latest version, also vary the name for the fun of it
     transform_v2 = ln.Transform(
-        name="Introduction v2", revises=transform_v1, version="2"
-    )
-    transform_v2.save()
+        description="Introduction v2", revises=transform_v1, version="2"
+    ).save()
     assert not transform_v1.is_latest
     assert transform_v2.is_latest
     assert transform_v2.uid.endswith("0001")
     # consciously *not* pass the latest version to revises but the previous
     # it automatically retrieves the latest version
-    transform_v3 = ln.Transform(name="Introduction", revises=transform_v1)
-    transform_v3.save()
+    transform_v3 = ln.Transform(description="Introduction", revises=transform_v1).save()
     assert transform_v3.uid.endswith("0002")
-    assert not ln.Transform.objects.get(name="Introduction v2", version="2").is_latest
+    assert not ln.Transform.objects.get(
+        description="Introduction v2", version="2"
+    ).is_latest
     assert transform_v3.is_latest
-    transform_v4 = ln.Transform(name="Introduction")
-    transform_v4.save()
+    transform_v4 = ln.Transform(description="Introduction").save()
     assert transform_v4.is_latest
     # add another transform with the same name that's not part of this family
     # but will also be a hit for the query
-    assert len(ln.Transform.filter(name="Introduction").all()) == 3
-    assert len(ln.Transform.filter(name="Introduction").latest_version()) == 2
+    assert len(ln.Transform.filter(description="Introduction").all()) == 3
+    assert len(ln.Transform.filter(description="Introduction").latest_version()) == 2
     transform_v4.delete()
     with pytest.raises(Exception):  # noqa: B017 should be MultipleResultsFound
-        ln.Transform.get(name="Introduction")
+        ln.Transform.get(description="Introduction")
     assert (
-        ln.Transform.filter(name="Introduction").latest_version().one() == transform_v3
+        ln.Transform.filter(description="Introduction").latest_version().one()
+        == transform_v3
     )
 
     # test get
@@ -132,6 +131,8 @@ def test_latest_version_and_get():
 
     # test empty QuerySet
     assert (
-        ln.Transform.filter(name="IntroductionNotExists").latest_version().one_or_none()
+        ln.Transform.filter(description="IntroductionNotExists")
+        .latest_version()
+        .one_or_none()
         is None
     )
